% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_helper.R
\docType{data}
\name{.system_prompt}
\alias{.system_prompt}
\title{Prompt for the LLM}
\format{
An object of class \code{character} of length 1.
}
\usage{
.system_prompt
}
\description{
Based on https://idekerlab.ucsd.edu/gsai/
}
\keyword{datasets}
