% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_shiny_treelabel.R
\name{init_shinyTreelabel}
\alias{init_shinyTreelabel}
\title{Prepare App}
\usage{
init_shinyTreelabel(
  data,
  treelabels = where(is_treelabel),
  design = ~1,
  pseudobulk_by = NULL,
  metaanalysis_over = NULL,
  gene_expr_by = NULL,
  contrasts = vars(Intercept = cond()),
  col_data = NULL,
  count_assay = "counts",
  seurat_assay = "RNA",
  treelabel_threshold = 0.5,
  reduced_dims = NULL
)
}
\description{
Prepare App
}
