% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_shiny_treelabel.R
\name{make_treelabel_app}
\alias{make_treelabel_app}
\title{Wrapper for shinyApp}
\usage{
make_treelabel_app(
  spec,
  sce = NULL,
  col_data = NULL,
  precalc_results = NULL,
  ...
)
}
\arguments{
\item{spec}{object generated by \code{init_shinyTreelabel} that describes the content that will be displayed}

\item{sce}{a \code{SingleCellExperiment}. It is used to calculate the differential expression results. If
\code{precalc_results} is provided and it is complete (containing DA and DE results), \code{sce} is not needed.}

\item{col_data}{a \code{data.frame}. It is used to calculate the differential abundance results. If
\code{precalc_results} is provided and it is complete (containing DA and DE results), \code{col_data} is not needed.}

\item{precalc_results}{the output from \code{precalculate_results} to avoid calculating expensive
results on the fly.}

\item{...}{the parameters forwarded to \code{shiny::shinyApp}}
}
\value{
Does not return anything, but blocks the thread and serves HTTP responses.
}
\description{
Wrapper for shinyApp
}
