% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_shiny_treelabel.R
\name{precalculate_results}
\alias{precalculate_results}
\alias{install_precalculated_results}
\title{Speed-up app by precalculating the results}
\usage{
precalculate_results(spec = c("all", "nothing"), verbose = TRUE)

install_precalculated_results(results)
}
\arguments{
\item{spec}{the specification which results to precalculate. Either a
string 'all' or 'nothing' or a list with two elements 'treelabels' and
'nodes' which list the elements to precalculate.}

\item{verbose}{should a progress bar be printed.}

\item{results}{a list produced by \code{precalculate_results}.}
}
\value{
a list with all the precalculated results that needs to be
passed to \code{install_precalculated_results}.
}
\description{
Speed-up app by precalculating the results
}
