% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_shiny_treelabel.R
\name{init_shinyTreelabel}
\alias{init_shinyTreelabel}
\title{Prepare App}
\usage{
init_shinyTreelabel(
  sce,
  treelabels = where(is_treelabel),
  design = ~1,
  pseudobulk_by = NULL,
  metaanalysis_over = NULL,
  gene_expr_by = NULL,
  contrasts = vars(Intercept = cond()),
  de_test = c("limma", "glmGamPoi"),
  dim_reductions = NULL,
  count_assay = "counts",
  treelabel_threshold = 0.5
)
}
\description{
Prepare App
}
