% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_shiny_treelabel.R
\name{precalculate_results}
\alias{precalculate_results}
\title{Speed-up app by precalculating the results}
\usage{
precalculate_results(
  spec,
  sce,
  output = c("da", "de", "de_meta"),
  treelabel_sel = NULL,
  dbfile = ":memory:",
  clear_file = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{spec}{the specification which results to precalculat generated by \code{init_shinyTreelabel}.}

\item{sce}{the \code{SingleCellExperiment} object that is used to calculate the differential abundance
and differential expression results.}

\item{output}{specify which set of results are calculate. By default differential abundance, differential
expression, and the meta analysis of the differential expression are calculated.}

\item{treelabel_sel}{calcualte the results only for a subset of the treelabel columns. Default: use all
from \code{spec$treelabel_names}.}

\item{dbfile}{the location where the precalculated results are stored. Default: \code{":memory:"} (i.e.,
in memory).}

\item{verbose}{should a progress bar be printed.}
}
\value{
an object of the R6 class \code{PrecalculatedResults}.
}
\description{
Speed-up app by precalculating the results
}
